requirements.txt
python-telegram-bot==20.7
Flask
Stripe
Supabase
python-dotenv

